Use `setuptools` and `setuptools-scm` to install the python bindings
`setuptools` (through `setuptools-scm`) supports automatically detecting
the tag and commit number of the git repo and use that as the version
number of the package. This commits implements this workflow, while also
keeping the version number stored in `defs.h` as a fallback value when
the source dir doesn't contain a `.git` folder.