Fix #71
The code will now check at the beginning whether a strand is shiftable or not based on the external forces acting on its nucleotides. If a strand is not shiftable then "fix_diffusion" will not act on it.
For now, the list that disable fix_diffusion on a per-strand basis are "trap", "twist" and "string".
Note that this change should make #73 unnecessary